Allow duplicate classes inside `org.junit.platform.commons.annotation.*`
The reason for this is that `kotlintest` and `junit` both share classes inside this package. This is fine when IntellIJ runs the tests, but the Maven Enforcer plugin complains (`banDuplicateClasses` rule). However, when the offending artifact is excluded from the dependency (satisfying Maven), the IntelliJ tests fail. This issue has been mitigated by simply telling Maven to ignore this violation.